+++ RELEASE_NOTES.md
@@ -1,2 +1,4 @@
 v0.67.4
 
+- Android and iOS: ChatConnectionListener has a new optional callback for server alerts.
+  (Already added for Node in v0.67.2.)